Update version + instructions for latest code
When using v3.4.0-231.0.dev or newer, debugging in VS Code works better (including breakpoints and stepping through the analyzer-provided code). This updates constraints to that version and tweaks the instructions a little (plus adds VS Code settings/launch config to avoid needing to be set manually to run the example).