Add a note about VS Code setting no longer being required (#5)
As of todays updates to the Dart extension, this setting isn't
necessary. I left it in with a note for now (since I don't know how
quickly people will update), but it could probably be dropped in the
near future to simplify things a little.

(Removing this setting avoids the case where macros are enabled in the
analysis_options, but the IDE support isn't enabled so things like
Go-to-Definition appear broken).